requirements.txt
pre-commit==3.2.2
fastapi==0.95.1
uvicorn[starlette]==0.21.1
openai==0.27.4
python-dotenv==1.0.0
llama-index==0.5.23.post1
langchain==0.0.142
